refactored tabs in profile page using antd
