+++ src/styles/App.css
@@ -1,10 +1,4 @@
-<<<<<<< HEAD
 
-.App {
-  text-align: center;
-}
-
-=======
 .App {
   text-align: center;
 }
@@ -33,15 +27,11 @@
   width: 1000px;
   height: 500px;
 }
->>>>>>> 7d978d3be8560753b6a2e9636b66d8b96a8c4ca3
+
 .App-logo {
   height: 40vmin;
 }
 
-<<<<<<< HEAD
-
-=======
->>>>>>> 7d978d3be8560753b6a2e9636b66d8b96a8c4ca3
 .App-link {
   color: #09d3ac;
 }
